requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
flatbuffers==24.3.25
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.9.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pyarrow==17.0.0
pydeck==0.9.1
pygame==2.6.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.39.0
streamlit-webrtc>=0.48
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.4.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
